requirements.txt
numpy==2.4.3
pandas==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
six==1.17.0
